avatar profil ok; updateinfos ok

+++ backend/02_user/controllers/controllers.js
@@ -80,7 +80,7 @@ export async function register(request, reply) {
 
     const user = await db.registered.getByName(name);
 	delete user.hashedPassword;
-	delete user.email;
+	//delete user.email;
 	delete user.telephone;
 
 	const { accessToken } = await sendCode({
@@ -125,7 +125,7 @@ export async function logIn(request, reply) {
 
         const user = await db.registered.getByName(name);
 		delete user.hashedPassword;
-		delete user.email;
+		//delete user.email;
 		
 		user.verified = true;
 
@@ -209,6 +209,7 @@ export async function updateInfo(request, reply) {
 		return reply.code(401).send( { error : 'User not Authentified'});
 		
 	const { password, toUpdate, newValue } = request.body;
+	console.log("toUpdate =", toUpdate, " newValue =", newValue);
 	if (!password || !toUpdate || !newValue)
 		return reply.code(401).send( { error : 'Need all infos in body'});
 
@@ -265,11 +266,14 @@ export async function updateAvatar(request, reply) {
 	if (!data)
 		return reply.code(400).send({ error: 'No file uploaded' });
 
+	if (data.filename.toLowerCase() === 'bg.webp')
+		return reply.code(400).send({ error: 'Invalid file name: BG.webp is reserved.' });
+
 	const uploadDir = path.join(process.cwd(), 'pictures');
 	if (!fs.existsSync(uploadDir))
 		fs.mkdirSync(uploadDir, { recursive: true });
 
-	if (user.picture && fs.existsSync(user.picture))
+	if (user.picture && fs.existsSync(user.picture) && !user.picture.endsWith('BG.webp'))
 		fs.unlinkSync(user.picture);
 
 	const ext = path.extname(data.filename);
@@ -279,16 +283,12 @@ export async function updateAvatar(request, reply) {
     const buffer = await data.toBuffer();
     fs.writeFileSync(filePath, buffer);
 
-    const relativePath = `/pictures/${fileName}`;
+    const relativePath = `pictures/${fileName}`;
     await db.registered.updateCol('picture', name, relativePath);
 
-    // Creation de  l'URL complete
-    const host = 'https://' + request.hostname + '4343';
-    const fullUrl = `/user/${host}${relativePath}`;
-	
 	return reply.code(200).send({
 		message: 'Avatar updated successfully',
-		picture: fullUrl
+		picture: relativePath
 	});
 }
 
@@ -471,7 +471,7 @@ export async function changeStatus(request, reply) {
 
 	delete user.hashedPassword;
 	delete user.telephone;
-	delete user.email;
+	//delete user.email;
 	return reply.code(201).send({
 		user: user,
 		message : 'Status updated!',
